+++ lib/modl/parser/MODLParser.rb
@@ -947,6 +947,14 @@ class MODLParser < Antlr4::Runtime::Parser
 		def modl_structure_i( i)
 			return rule_context("Modl_structureContext",i)
 		end
+
+    def STRUCT_SEP()
+      return tokens(MODLParser::STRUCT_SEP)
+    end
+
+    def STRUCT_SEP_i(i)
+      return token(MODLParser::STRUCT_SEP, i)
+    end
 		def initialize( parent,  invokingState)
 			super(parent, invokingState)
 		end
@@ -980,15 +988,19 @@ class MODLParser < Antlr4::Runtime::Parser
 		begin
 			enter_outer_alt(_localctx, 1)
 
-			@_state_number = 166
+      @_state_number = 163
+      modl_structure()
+      @_state_number = 168
 			@_err_handler.sync(self)
 			_la = @_input.la(1)
-      while ((((_la) & ~0x3f) == 0 && ((1 << _la) & ((1 << LBRAC) | (1 << LSBRAC) | (1 << QUOTED) | (1 << STRING) | (1 << LCBRAC))) != 0))
+      while (_la == STRUCT_SEP)
 
 
-				@_state_number = 163
+        @_state_number = 164
+        match(STRUCT_SEP)
+        @_state_number = 165
 				modl_structure()
-				@_state_number = 168
+        @_state_number = 170
 				@_err_handler.sync(self)
 				_la = @_input.la(1)
       end
@@ -1066,40 +1078,40 @@ class MODLParser < Antlr4::Runtime::Parser
 		begin
 			enter_outer_alt(_localctx, 1)
 
-			@_state_number = 169
+      @_state_number = 171
 			match(LCBRAC)
-			@_state_number = 170
+      @_state_number = 172
 			modl_condition_test()
-			@_state_number = 171
+      @_state_number = 173
 			match(QMARK)
-			@_state_number = 172
+      @_state_number = 174
 			modl_map_conditional_return()
-			@_state_number = 181
+      @_state_number = 183
 			@_err_handler.sync(self)
 			_la = @_input.la(1)
 			while (_la==FSLASH)
 
 
-				@_state_number = 173
+        @_state_number = 175
 				match(FSLASH)
-				@_state_number = 175
+        @_state_number = 177
 				@_err_handler.sync(self)
 				_la = @_input.la(1)
         if ((((_la) & ~0x3f) == 0 && ((1 << _la) & ((1 << NULL) | (1 << TRUE) | (1 << FALSE) | (1 << EQUALS) | (1 << LBRAC) | (1 << LSBRAC) | (1 << NUMBER) | (1 << QUOTED) | (1 << STRING) | (1 << LCBRAC) | (1 << GTHAN) | (1 << LTHAN) | (1 << EXCLAM))) != 0))
 
-					@_state_number = 174
+          @_state_number = 176
 					modl_condition_test()
         end
 
-				@_state_number = 177
+        @_state_number = 179
 				match(QMARK)
-				@_state_number = 178
+        @_state_number = 180
 				modl_map_conditional_return()
-				@_state_number = 183
+        @_state_number = 185
 				@_err_handler.sync(self)
 				_la = @_input.la(1)
 			end
-			@_state_number = 184
+      @_state_number = 186
 			match(RCBRAC)
 		rescue Antlr4::Runtime::RecognitionException => re
 			_localctx.exception = re
@@ -1151,16 +1163,16 @@ class MODLParser < Antlr4::Runtime::Parser
 		begin
 			enter_outer_alt(_localctx, 1)
 
-			@_state_number = 187 
-			@_err_handler.sync(self)
+      @_state_number = 189
+      @_err_handler.sync(self)
 			_la = @_input.la(1)
 			loop do
 
 
-				@_state_number = 186
+        @_state_number = 188
 				modl_map_item()
-				@_state_number = 189 
-				@_err_handler.sync(self)
+        @_state_number = 191
+        @_err_handler.sync(self)
 				_la = @_input.la(1)
         break if (!((((_la) & ~0x3f) == 0 && ((1 << _la) & ((1 << QUOTED) | (1 << STRING) | (1 << LCBRAC))) != 0)))
 			end
@@ -1211,19 +1223,19 @@ class MODLParser < Antlr4::Runtime::Parser
 		_localctx =  Modl_map_itemContext.new(@_ctx, @_state_number)
 		enter_rule(_localctx, 22, RULE_modl_map_item)
 		begin
-			@_state_number = 193
+      @_state_number = 195
 			@_err_handler.sync(self)
       case (@_input.la(1))
       when MODLParser::QUOTED, MODLParser::STRING
 				enter_outer_alt(_localctx, 1)
 
-				@_state_number = 191
+        @_state_number = 193
 				modl_pair()
 
       when MODLParser::LCBRAC
 				enter_outer_alt(_localctx, 2)
 
-				@_state_number = 192
+        @_state_number = 194
 				modl_map_conditional()
       else
 				raise Antlr4::Runtime::NoViableAltException, self
@@ -1302,40 +1314,40 @@ class MODLParser < Antlr4::Runtime::Parser
 		begin
 			enter_outer_alt(_localctx, 1)
 
-			@_state_number = 195
+      @_state_number = 197
 			match(LCBRAC)
-			@_state_number = 196
+      @_state_number = 198
 			modl_condition_test()
-			@_state_number = 197
+      @_state_number = 199
 			match(QMARK)
-			@_state_number = 198
+      @_state_number = 200
 			modl_array_conditional_return()
-			@_state_number = 207
+      @_state_number = 209
 			@_err_handler.sync(self)
 			_la = @_input.la(1)
 			while (_la==FSLASH)
 
 
-				@_state_number = 199
+        @_state_number = 201
 				match(FSLASH)
-				@_state_number = 201
+        @_state_number = 203
 				@_err_handler.sync(self)
 				_la = @_input.la(1)
         if ((((_la) & ~0x3f) == 0 && ((1 << _la) & ((1 << NULL) | (1 << TRUE) | (1 << FALSE) | (1 << EQUALS) | (1 << LBRAC) | (1 << LSBRAC) | (1 << NUMBER) | (1 << QUOTED) | (1 << STRING) | (1 << LCBRAC) | (1 << GTHAN) | (1 << LTHAN) | (1 << EXCLAM))) != 0))
 
-					@_state_number = 200
+          @_state_number = 202
 					modl_condition_test()
         end
 
-				@_state_number = 203
+        @_state_number = 205
 				match(QMARK)
-				@_state_number = 204
+        @_state_number = 206
 				modl_array_conditional_return()
-				@_state_number = 209
+        @_state_number = 211
 				@_err_handler.sync(self)
 				_la = @_input.la(1)
 			end
-			@_state_number = 210
+      @_state_number = 212
 			match(RCBRAC)
 		rescue Antlr4::Runtime::RecognitionException => re
 			_localctx.exception = re
@@ -1387,16 +1399,16 @@ class MODLParser < Antlr4::Runtime::Parser
 		begin
 			enter_outer_alt(_localctx, 1)
 
-			@_state_number = 213 
-			@_err_handler.sync(self)
+      @_state_number = 215
+      @_err_handler.sync(self)
 			_la = @_input.la(1)
 			loop do
 
 
-				@_state_number = 212
+        @_state_number = 214
 				modl_array_item()
-				@_state_number = 215 
-				@_err_handler.sync(self)
+        @_state_number = 217
+        @_err_handler.sync(self)
 				_la = @_input.la(1)
         break if (!((((_la) & ~0x3f) == 0 && ((1 << _la) & ((1 << NULL) | (1 << TRUE) | (1 << FALSE) | (1 << LBRAC) | (1 << LSBRAC) | (1 << NUMBER) | (1 << QUOTED) | (1 << STRING) | (1 << LCBRAC))) != 0)))
 			end
@@ -1447,19 +1459,19 @@ class MODLParser < Antlr4::Runtime::Parser
 		_localctx =  Modl_array_itemContext.new(@_ctx, @_state_number)
 		enter_rule(_localctx, 28, RULE_modl_array_item)
 		begin
-			@_state_number = 219
+      @_state_number = 221
 			@_err_handler.sync(self)
       case (@_input.la(1))
       when MODLParser::NULL, MODLParser::TRUE, MODLParser::FALSE, MODLParser::LBRAC, MODLParser::LSBRAC, MODLParser::NUMBER, MODLParser::QUOTED, MODLParser::STRING
 				enter_outer_alt(_localctx, 1)
 
-				@_state_number = 217
+        @_state_number = 219
 				modl_array_value_item()
 
       when MODLParser::LCBRAC
 				enter_outer_alt(_localctx, 2)
 
-				@_state_number = 218
+        @_state_number = 220
 				modl_array_conditional()
       else
 				raise Antlr4::Runtime::NoViableAltException, self
@@ -1539,49 +1551,49 @@ class MODLParser < Antlr4::Runtime::Parser
 			_alt = 0
 			enter_outer_alt(_localctx, 1)
 
-			@_state_number = 221
+      @_state_number = 223
 			match(LCBRAC)
-			@_state_number = 222
+      @_state_number = 224
 			modl_condition_test()
-			@_state_number = 223
+      @_state_number = 225
 			match(QMARK)
-			@_state_number = 239
+      @_state_number = 241
 			@_err_handler.sync(self)
 			_la = @_input.la(1)
       if ((((_la) & ~0x3f) == 0 && ((1 << _la) & ((1 << NULL) | (1 << TRUE) | (1 << FALSE) | (1 << LBRAC) | (1 << LSBRAC) | (1 << NUMBER) | (1 << QUOTED) | (1 << STRING) | (1 << LCBRAC))) != 0))
 
-				@_state_number = 224
+        @_state_number = 226
 				modl_value_conditional_return()
-				@_state_number = 232
+        @_state_number = 234
 				@_err_handler.sync(self);
 				_alt = @_interp.adaptive_predict(@_input,30,@_ctx)
 				while ( _alt!=2 && _alt!=Antlr4::Runtime::ATN::INVALID_ALT_NUMBER )
 					if ( _alt==1 )
 
 
-						@_state_number = 225
+            @_state_number = 227
 						match(FSLASH)
-						@_state_number = 226
+            @_state_number = 228
 						modl_condition_test()
-						@_state_number = 227
+            @_state_number = 229
 						match(QMARK)
-						@_state_number = 228
+            @_state_number = 230
 						modl_value_conditional_return() 
 					end
-					@_state_number = 234
+          @_state_number = 236
 					@_err_handler.sync(self)
 					_alt = @_interp.adaptive_predict(@_input,30,@_ctx)
 				end
 
-				@_state_number = 235
+        @_state_number = 237
 				match(FSLASH)
-				@_state_number = 236
+        @_state_number = 238
 				match(QMARK)
-				@_state_number = 237
+        @_state_number = 239
 				modl_value_conditional_return()
       end
 
-			@_state_number = 241
+      @_state_number = 243
 			match(RCBRAC)
 		rescue Antlr4::Runtime::RecognitionException => re
 			_localctx.exception = re
@@ -1639,19 +1651,19 @@ class MODLParser < Antlr4::Runtime::Parser
 		begin
 			enter_outer_alt(_localctx, 1)
 
-			@_state_number = 243
+      @_state_number = 245
 			modl_value_item()
-			@_state_number = 248
+      @_state_number = 250
 			@_err_handler.sync(self)
 			_la = @_input.la(1)
 			while (_la==COLON)
 
 
-				@_state_number = 244
+        @_state_number = 246
 				match(COLON)
-				@_state_number = 245
+        @_state_number = 247
 				modl_value_item()
-				@_state_number = 250
+        @_state_number = 252
 				@_err_handler.sync(self)
 				_la = @_input.la(1)
 			end
@@ -1730,37 +1742,37 @@ class MODLParser < Antlr4::Runtime::Parser
 			_alt = 0
 			enter_outer_alt(_localctx, 1)
 
-			@_state_number = 252
+      @_state_number = 254
 			@_err_handler.sync(self)
 			case ( @_interp.adaptive_predict(@_input,33,@_ctx) )
 			when 1
 
-				@_state_number = 251
+        @_state_number = 253
 				match(EXCLAM)
 
 			end
-			@_state_number = 256
+      @_state_number = 258
 			@_err_handler.sync(self)
 			case ( @_interp.adaptive_predict(@_input,34,@_ctx) )
 			when 1
 
-				@_state_number = 254
+        @_state_number = 256
 				modl_condition()
 
 			when 2
 
-				@_state_number = 255
+        @_state_number = 257
 				modl_condition_group()
 
 			end
-			@_state_number = 268
+      @_state_number = 270
 			@_err_handler.sync(self);
 			_alt = @_interp.adaptive_predict(@_input,37,@_ctx)
 			while ( _alt!=2 && _alt!=Antlr4::Runtime::ATN::INVALID_ALT_NUMBER )
 				if ( _alt==1 )
 
 
-					@_state_number = 258
+          @_state_number = 260
 					_la = @_input.la(1)
 					if ( !(_la==AMP || _la==PIPE) )
 					@_err_handler.recover_in_line(self)
@@ -1772,31 +1784,31 @@ class MODLParser < Antlr4::Runtime::Parser
 						@_err_handler.report_match(self)
 						consume()
 					end
-					@_state_number = 260
+          @_state_number = 262
 					@_err_handler.sync(self)
 					case ( @_interp.adaptive_predict(@_input,35,@_ctx) )
 					when 1
 
-						@_state_number = 259
+            @_state_number = 261
 						match(EXCLAM)
 
 					end
-					@_state_number = 264
+          @_state_number = 266
 					@_err_handler.sync(self)
 					case ( @_interp.adaptive_predict(@_input,36,@_ctx) )
 					when 1
 
-						@_state_number = 262
+            @_state_number = 264
 						modl_condition()
 
 					when 2
 
-						@_state_number = 263
+            @_state_number = 265
 						modl_condition_group()
 
 					end 
 				end
-				@_state_number = 270
+        @_state_number = 272
 				@_err_handler.sync(self)
 				_alt = @_interp.adaptive_predict(@_input,37,@_ctx)
 			end
@@ -1853,49 +1865,49 @@ class MODLParser < Antlr4::Runtime::Parser
 		_localctx =  Modl_operatorContext.new(@_ctx, @_state_number)
 		enter_rule(_localctx, 36, RULE_modl_operator)
 		begin
-			@_state_number = 280
+      @_state_number = 282
 			@_err_handler.sync(self)
 			case ( @_interp.adaptive_predict(@_input,38,@_ctx) )
 			when 1
 				enter_outer_alt(_localctx, 1)
 
-				@_state_number = 271
+        @_state_number = 273
 				match(EQUALS)
 
 			when 2
 				enter_outer_alt(_localctx, 2)
 
-				@_state_number = 272
+        @_state_number = 274
 				match(GTHAN)
 
 			when 3
 				enter_outer_alt(_localctx, 3)
 
-				@_state_number = 273
+        @_state_number = 275
 				match(GTHAN)
-				@_state_number = 274
+        @_state_number = 276
 				match(EQUALS)
 
 			when 4
 				enter_outer_alt(_localctx, 4)
 
-				@_state_number = 275
+        @_state_number = 277
 				match(LTHAN)
 
 			when 5
 				enter_outer_alt(_localctx, 5)
 
-				@_state_number = 276
+        @_state_number = 278
 				match(LTHAN)
-				@_state_number = 277
+        @_state_number = 279
 				match(EQUALS)
 
 			when 6
 				enter_outer_alt(_localctx, 6)
 
-				@_state_number = 278
+        @_state_number = 280
 				match(EXCLAM)
-				@_state_number = 279
+        @_state_number = 281
 				match(EQUALS)
 
 			end
@@ -1961,37 +1973,37 @@ class MODLParser < Antlr4::Runtime::Parser
 		begin
 			enter_outer_alt(_localctx, 1)
 
-			@_state_number = 283
+      @_state_number = 285
 			@_err_handler.sync(self)
 			case ( @_interp.adaptive_predict(@_input,39,@_ctx) )
 			when 1
 
-				@_state_number = 282
+        @_state_number = 284
 				match(STRING)
 
 			end
-			@_state_number = 286
+      @_state_number = 288
 			@_err_handler.sync(self)
 			_la = @_input.la(1)
 			if ((((_la) & ~0x3f) == 0 && ((1 << _la) & ((1 << EQUALS) | (1 << GTHAN) | (1 << LTHAN) | (1 << EXCLAM))) != 0))
 
-				@_state_number = 285
+        @_state_number = 287
 				modl_operator()
 			end
 
-			@_state_number = 288
+      @_state_number = 290
 			modl_value()
-			@_state_number = 293
+      @_state_number = 295
 			@_err_handler.sync(self)
 			_la = @_input.la(1)
 			while (_la==FSLASH)
 
 
-				@_state_number = 289
+        @_state_number = 291
 				match(FSLASH)
-				@_state_number = 290
+        @_state_number = 292
 				modl_value()
-				@_state_number = 295
+        @_state_number = 297
 				@_err_handler.sync(self)
 				_la = @_input.la(1)
 			end
@@ -2063,17 +2075,17 @@ class MODLParser < Antlr4::Runtime::Parser
 		begin
 			enter_outer_alt(_localctx, 1)
 
-			@_state_number = 296
+      @_state_number = 298
 			match(LCBRAC)
-			@_state_number = 297
+      @_state_number = 299
 			modl_condition_test()
-			@_state_number = 302
+      @_state_number = 304
 			@_err_handler.sync(self)
 			_la = @_input.la(1)
 			while (_la==AMP || _la==PIPE)
 
 
-				@_state_number = 298
+        @_state_number = 300
 				_la = @_input.la(1)
 				if ( !(_la==AMP || _la==PIPE) )
 				@_err_handler.recover_in_line(self)
@@ -2085,13 +2097,13 @@ class MODLParser < Antlr4::Runtime::Parser
 					@_err_handler.report_match(self)
 					consume()
 				end
-				@_state_number = 299
+        @_state_number = 301
 				modl_condition_test()
-				@_state_number = 304
+        @_state_number = 306
 				@_err_handler.sync(self)
 				_la = @_input.la(1)
 			end
-			@_state_number = 305
+      @_state_number = 307
 			match(RCBRAC)
 		rescue Antlr4::Runtime::RecognitionException => re
 			_localctx.exception = re
@@ -2149,37 +2161,37 @@ class MODLParser < Antlr4::Runtime::Parser
 		_localctx =  Modl_valueContext.new(@_ctx, @_state_number)
 		enter_rule(_localctx, 42, RULE_modl_value)
 		begin
-			@_state_number = 312
+      @_state_number = 314
 			@_err_handler.sync(self)
 			case ( @_interp.adaptive_predict(@_input,43,@_ctx) )
 			when 1
 				enter_outer_alt(_localctx, 1)
 
-				@_state_number = 307
+        @_state_number = 309
 				modl_map()
 
 			when 2
 				enter_outer_alt(_localctx, 2)
 
-				@_state_number = 308
+        @_state_number = 310
 				modl_array()
 
 			when 3
 				enter_outer_alt(_localctx, 3)
 
-				@_state_number = 309
+        @_state_number = 311
 				modl_nb_array()
 
 			when 4
 				enter_outer_alt(_localctx, 4)
 
-				@_state_number = 310
+        @_state_number = 312
 				modl_pair()
 
 			when 5
 				enter_outer_alt(_localctx, 5)
 
-				@_state_number = 311
+        @_state_number = 313
 				modl_primitive()
 
 			end
@@ -2236,31 +2248,31 @@ class MODLParser < Antlr4::Runtime::Parser
 		_localctx =  Modl_array_value_itemContext.new(@_ctx, @_state_number)
 		enter_rule(_localctx, 44, RULE_modl_array_value_item)
 		begin
-			@_state_number = 318
+      @_state_number = 320
 			@_err_handler.sync(self)
 			case ( @_interp.adaptive_predict(@_input,44,@_ctx) )
 			when 1
 				enter_outer_alt(_localctx, 1)
 
-				@_state_number = 314
+        @_state_number = 316
 				modl_map()
 
 			when 2
 				enter_outer_alt(_localctx, 2)
 
-				@_state_number = 315
+        @_state_number = 317
 				modl_pair()
 
 			when 3
 				enter_outer_alt(_localctx, 3)
 
-				@_state_number = 316
+        @_state_number = 318
 				modl_array()
 
 			when 4
 				enter_outer_alt(_localctx, 4)
 
-				@_state_number = 317
+        @_state_number = 319
 				modl_primitive()
 
 			end
@@ -2326,7 +2338,7 @@ class MODLParser < Antlr4::Runtime::Parser
 		begin
 			enter_outer_alt(_localctx, 1)
 
-			@_state_number = 320
+      @_state_number = 322
 			_la = @_input.la(1)
       if (!((((_la) & ~0x3f) == 0 && ((1 << _la) & ((1 << NULL) | (1 << TRUE) | (1 << FALSE) | (1 << NUMBER) | (1 << QUOTED) | (1 << STRING))) != 0)))
 			@_err_handler.recover_in_line(self)
@@ -2349,7 +2361,7 @@ class MODLParser < Antlr4::Runtime::Parser
 	end
 
 	@@_serializedATN = ["\3\u608b\ua72a\u8133\ub9ed\u417c\u3be7\u7786\u5964",
-                      "\3\37\u0145\4\2\t\2\4\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7\t\7\4\b\t\b",
+                      "\3\37\u0147\4\2\t\2\4\3\t\3\4\4\t\4\4\5\t\5\4\6\t\6\4\7\t\7\4\b\t\b",
                       "\4\t\t\t\4\n\t\n\4\13\t\13\4\f\t\f\4\r\t\r\4\16\t\16\4\17\t\17\4\20",
                       "\t\20\4\21\t\21\4\22\t\22\4\23\t\23\4\24\t\24\4\25\t\25\4\26\t\26\4",
                       "\27\t\27\4\30\t\30\4\31\t\31\3\2\5\2\64\n\2\3\2\3\2\3\2\7\29\n\2\f",
@@ -2360,29 +2372,29 @@ class MODLParser < Antlr4::Runtime::Parser
                       "\6w\n\6\r\6\16\6x\6\6{\n\6\r\6\16\6|\3\6\7\6\u0080\n\6\f\6\16\6\u0083",
                       "\13\6\3\6\5\6\u0086\n\6\3\7\3\7\3\7\3\7\3\7\3\7\3\7\5\7\u008f\n\7\3",
                       "\b\3\b\5\b\u0093\n\b\3\t\3\t\3\t\3\t\3\t\3\t\5\t\u009b\n\t\3\t\3\t",
-                      "\7\t\u009f\n\t\f\t\16\t\u00a2\13\t\3\t\3\t\3\n\7\n\u00a7\n\n\f\n\16",
-                      "\n\u00aa\13\n\3\13\3\13\3\13\3\13\3\13\3\13\5\13\u00b2\n\13\3\13\3",
-                      "\13\7\13\u00b6\n\13\f\13\16\13\u00b9\13\13\3\13\3\13\3\f\6\f\u00be",
-                      "\n\f\r\f\16\f\u00bf\3\r\3\r\5\r\u00c4\n\r\3\16\3\16\3\16\3\16\3\16",
-                      "\3\16\5\16\u00cc\n\16\3\16\3\16\7\16\u00d0\n\16\f\16\16\16\u00d3\13",
-                      "\16\3\16\3\16\3\17\6\17\u00d8\n\17\r\17\16\17\u00d9\3\20\3\20\5\20",
-                      "\u00de\n\20\3\21\3\21\3\21\3\21\3\21\3\21\3\21\3\21\3\21\7\21\u00e9",
-                      "\n\21\f\21\16\21\u00ec\13\21\3\21\3\21\3\21\3\21\5\21\u00f2\n\21\3",
-                      "\21\3\21\3\22\3\22\3\22\7\22\u00f9\n\22\f\22\16\22\u00fc\13\22\3\23",
-                      "\5\23\u00ff\n\23\3\23\3\23\5\23\u0103\n\23\3\23\3\23\5\23\u0107\n\23",
-                      "\3\23\3\23\5\23\u010b\n\23\7\23\u010d\n\23\f\23\16\23\u0110\13\23\3",
-                      "\24\3\24\3\24\3\24\3\24\3\24\3\24\3\24\3\24\5\24\u011b\n\24\3\25\5",
-                      "\25\u011e\n\25\3\25\5\25\u0121\n\25\3\25\3\25\3\25\7\25\u0126\n\25",
-                      "\f\25\16\25\u0129\13\25\3\26\3\26\3\26\3\26\7\26\u012f\n\26\f\26\16",
-                      "\26\u0132\13\26\3\26\3\26\3\27\3\27\3\27\3\27\3\27\5\27\u013b\n\27",
-                      "\3\30\3\30\3\30\3\30\5\30\u0141\n\30\3\31\3\31\3\31\2\2\32\2\4\6\b",
+                      "\7\t\u009f\n\t\f\t\16\t\u00a2\13\t\3\t\3\t\3\n\3\n\3\n\7\n\u00a9\n",
+                      "\n\f\n\16\n\u00ac\13\n\3\13\3\13\3\13\3\13\3\13\3\13\5\13\u00b4\n\13",
+                      "\3\13\3\13\7\13\u00b8\n\13\f\13\16\13\u00bb\13\13\3\13\3\13\3\f\6\f",
+                      "\u00c0\n\f\r\f\16\f\u00c1\3\r\3\r\5\r\u00c6\n\r\3\16\3\16\3\16\3\16",
+                      "\3\16\3\16\5\16\u00ce\n\16\3\16\3\16\7\16\u00d2\n\16\f\16\16\16\u00d5",
+                      "\13\16\3\16\3\16\3\17\6\17\u00da\n\17\r\17\16\17\u00db\3\20\3\20\5",
+                      "\20\u00e0\n\20\3\21\3\21\3\21\3\21\3\21\3\21\3\21\3\21\3\21\7\21\u00eb",
+                      "\n\21\f\21\16\21\u00ee\13\21\3\21\3\21\3\21\3\21\5\21\u00f4\n\21\3",
+                      "\21\3\21\3\22\3\22\3\22\7\22\u00fb\n\22\f\22\16\22\u00fe\13\22\3\23",
+                      "\5\23\u0101\n\23\3\23\3\23\5\23\u0105\n\23\3\23\3\23\5\23\u0109\n\23",
+                      "\3\23\3\23\5\23\u010d\n\23\7\23\u010f\n\23\f\23\16\23\u0112\13\23\3",
+                      "\24\3\24\3\24\3\24\3\24\3\24\3\24\3\24\3\24\5\24\u011d\n\24\3\25\5",
+                      "\25\u0120\n\25\3\25\5\25\u0123\n\25\3\25\3\25\3\25\7\25\u0128\n\25",
+                      "\f\25\16\25\u012b\13\25\3\26\3\26\3\26\3\26\7\26\u0131\n\26\f\26\16",
+                      "\26\u0134\13\26\3\26\3\26\3\27\3\27\3\27\3\27\3\27\5\27\u013d\n\27",
+                      "\3\30\3\30\3\30\3\30\5\30\u0143\n\30\3\31\3\31\3\31\2\2\32\2\4\6\b",
                       "\n\f\16\20\22\24\26\30\32\34\36 \"$&(*,.\60\2\5\3\2\21\22\3\2\33\34",
-                      "\5\2\4\6\17\17\21\22\2\u0165\2@\3\2\2\2\4H\3\2\2\2\6J\3\2\2\2\bW\3",
+                      "\5\2\4\6\17\17\21\22\2\u0167\2@\3\2\2\2\4H\3\2\2\2\6J\3\2\2\2\bW\3",
                       "\2\2\2\nz\3\2\2\2\f\u008e\3\2\2\2\16\u0092\3\2\2\2\20\u0094\3\2\2\2",
-                      "\22\u00a8\3\2\2\2\24\u00ab\3\2\2\2\26\u00bd\3\2\2\2\30\u00c3\3\2\2",
-                      "\2\32\u00c5\3\2\2\2\34\u00d7\3\2\2\2\36\u00dd\3\2\2\2 \u00df\3\2\2",
-                      "\2\"\u00f5\3\2\2\2$\u00fe\3\2\2\2&\u011a\3\2\2\2(\u011d\3\2\2\2*\u012a",
-                      "\3\2\2\2,\u013a\3\2\2\2.\u0140\3\2\2\2\60\u0142\3\2\2\2\62\64\5\4\3",
+                      "\22\u00a5\3\2\2\2\24\u00ad\3\2\2\2\26\u00bf\3\2\2\2\30\u00c5\3\2\2",
+                      "\2\32\u00c7\3\2\2\2\34\u00d9\3\2\2\2\36\u00df\3\2\2\2 \u00e1\3\2\2",
+                      "\2\"\u00f7\3\2\2\2$\u0100\3\2\2\2&\u011c\3\2\2\2(\u011f\3\2\2\2*\u012c",
+                      "\3\2\2\2,\u013c\3\2\2\2.\u0142\3\2\2\2\60\u0144\3\2\2\2\62\64\5\4\3",
                       "\2\63\62\3\2\2\2\63\64\3\2\2\2\64A\3\2\2\2\65:\5\4\3\2\66\67\7\t\2",
                       "\2\679\5\4\3\28\66\3\2\2\29<\3\2\2\2:8\3\2\2\2:;\3\2\2\2;>\3\2\2\2",
                       "<:\3\2\2\2=?\7\t\2\2>=\3\2\2\2>?\3\2\2\2?A\3\2\2\2@\63\3\2\2\2@\65",
@@ -2411,66 +2423,67 @@ class MODLParser < Antlr4::Runtime::Parser
                       "\2\u009c\u009d\7\26\2\2\u009d\u009f\5\22\n\2\u009e\u0098\3\2\2\2\u009f",
                       "\u00a2\3\2\2\2\u00a0\u009e\3\2\2\2\u00a0\u00a1\3\2\2\2\u00a1\u00a3",
                       "\3\2\2\2\u00a2\u00a0\3\2\2\2\u00a3\u00a4\7\37\2\2\u00a4\21\3\2\2\2",
-                      "\u00a5\u00a7\5\4\3\2\u00a6\u00a5\3\2\2\2\u00a7\u00aa\3\2\2\2\u00a8",
-                      "\u00a6\3\2\2\2\u00a8\u00a9\3\2\2\2\u00a9\23\3\2\2\2\u00aa\u00a8\3\2",
-                      "\2\2\u00ab\u00ac\7\24\2\2\u00ac\u00ad\5$\23\2\u00ad\u00ae\7\26\2\2",
-                      "\u00ae\u00b7\5\26\f\2\u00af\u00b1\7\27\2\2\u00b0\u00b2\5$\23\2\u00b1",
-                      "\u00b0\3\2\2\2\u00b1\u00b2\3\2\2\2\u00b2\u00b3\3\2\2\2\u00b3\u00b4",
-                      "\7\26\2\2\u00b4\u00b6\5\26\f\2\u00b5\u00af\3\2\2\2\u00b6\u00b9\3\2",
-                      "\2\2\u00b7\u00b5\3\2\2\2\u00b7\u00b8\3\2\2\2\u00b8\u00ba\3\2\2\2\u00b9",
-                      "\u00b7\3\2\2\2\u00ba\u00bb\7\37\2\2\u00bb\25\3\2\2\2\u00bc\u00be\5",
-                      "\30\r\2\u00bd\u00bc\3\2\2\2\u00be\u00bf\3\2\2\2\u00bf\u00bd\3\2\2\2",
-                      "\u00bf\u00c0\3\2\2\2\u00c0\27\3\2\2\2\u00c1\u00c4\5\f\7\2\u00c2\u00c4",
-                      "\5\24\13\2\u00c3\u00c1\3\2\2\2\u00c3\u00c2\3\2\2\2\u00c4\31\3\2\2\2",
-                      "\u00c5\u00c6\7\24\2\2\u00c6\u00c7\5$\23\2\u00c7\u00c8\7\26\2\2\u00c8",
-                      "\u00d1\5\34\17\2\u00c9\u00cb\7\27\2\2\u00ca\u00cc\5$\23\2\u00cb\u00ca",
-                      "\3\2\2\2\u00cb\u00cc\3\2\2\2\u00cc\u00cd\3\2\2\2\u00cd\u00ce\7\26\2",
-                      "\2\u00ce\u00d0\5\34\17\2\u00cf\u00c9\3\2\2\2\u00d0\u00d3\3\2\2\2\u00d1",
-                      "\u00cf\3\2\2\2\u00d1\u00d2\3\2\2\2\u00d2\u00d4\3\2\2\2\u00d3\u00d1",
-                      "\3\2\2\2\u00d4\u00d5\7\37\2\2\u00d5\33\3\2\2\2\u00d6\u00d8\5\36\20",
-                      "\2\u00d7\u00d6\3\2\2\2\u00d8\u00d9\3\2\2\2\u00d9\u00d7\3\2\2\2\u00d9",
-                      "\u00da\3\2\2\2\u00da\35\3\2\2\2\u00db\u00de\5.\30\2\u00dc\u00de\5\32",
-                      "\16\2\u00dd\u00db\3\2\2\2\u00dd\u00dc\3\2\2\2\u00de\37\3\2\2\2\u00df",
-                      "\u00e0\7\24\2\2\u00e0\u00e1\5$\23\2\u00e1\u00f1\7\26\2\2\u00e2\u00ea",
-                      "\5\"\22\2\u00e3\u00e4\7\27\2\2\u00e4\u00e5\5$\23\2\u00e5\u00e6\7\26",
-                      "\2\2\u00e6\u00e7\5\"\22\2\u00e7\u00e9\3\2\2\2\u00e8\u00e3\3\2\2\2\u00e9",
-                      "\u00ec\3\2\2\2\u00ea\u00e8\3\2\2\2\u00ea\u00eb\3\2\2\2\u00eb\u00ed",
-                      "\3\2\2\2\u00ec\u00ea\3\2\2\2\u00ed\u00ee\7\27\2\2\u00ee\u00ef\7\26",
-                      "\2\2\u00ef\u00f0\5\"\22\2\u00f0\u00f2\3\2\2\2\u00f1\u00e2\3\2\2\2\u00f1",
-                      "\u00f2\3\2\2\2\u00f2\u00f3\3\2\2\2\u00f3\u00f4\7\37\2\2\u00f4!\3\2",
-                      "\2\2\u00f5\u00fa\5\16\b\2\u00f6\u00f7\7\7\2\2\u00f7\u00f9\5\16\b\2",
-                      "\u00f8\u00f6\3\2\2\2\u00f9\u00fc\3\2\2\2\u00fa\u00f8\3\2\2\2\u00fa",
-                      "\u00fb\3\2\2\2\u00fb#\3\2\2\2\u00fc\u00fa\3\2\2\2\u00fd\u00ff\7\35",
-                      "\2\2\u00fe\u00fd\3\2\2\2\u00fe\u00ff\3\2\2\2\u00ff\u0102\3\2\2\2\u0100",
-                      "\u0103\5(\25\2\u0101\u0103\5*\26\2\u0102\u0100\3\2\2\2\u0102\u0101",
-                      "\3\2\2\2\u0103\u010e\3\2\2\2\u0104\u0106\t\3\2\2\u0105\u0107\7\35\2",
-                      "\2\u0106\u0105\3\2\2\2\u0106\u0107\3\2\2\2\u0107\u010a\3\2\2\2\u0108",
-                      "\u010b\5(\25\2\u0109\u010b\5*\26\2\u010a\u0108\3\2\2\2\u010a\u0109",
-                      "\3\2\2\2\u010b\u010d\3\2\2\2\u010c\u0104\3\2\2\2\u010d\u0110\3\2\2",
-                      "\2\u010e\u010c\3\2\2\2\u010e\u010f\3\2\2\2\u010f%\3\2\2\2\u0110\u010e",
-                      "\3\2\2\2\u0111\u011b\7\b\2\2\u0112\u011b\7\30\2\2\u0113\u0114\7\30",
-                      "\2\2\u0114\u011b\7\b\2\2\u0115\u011b\7\31\2\2\u0116\u0117\7\31\2\2",
-                      "\u0117\u011b\7\b\2\2\u0118\u0119\7\35\2\2\u0119\u011b\7\b\2\2\u011a",
-                      "\u0111\3\2\2\2\u011a\u0112\3\2\2\2\u011a\u0113\3\2\2\2\u011a\u0115",
-                      "\3\2\2\2\u011a\u0116\3\2\2\2\u011a\u0118\3\2\2\2\u011b\'\3\2\2\2\u011c",
-                      "\u011e\7\22\2\2\u011d\u011c\3\2\2\2\u011d\u011e\3\2\2\2\u011e\u0120",
-                      "\3\2\2\2\u011f\u0121\5&\24\2\u0120\u011f\3\2\2\2\u0120\u0121\3\2\2",
-                      "\2\u0121\u0122\3\2\2\2\u0122\u0127\5,\27\2\u0123\u0124\7\27\2\2\u0124",
-                      "\u0126\5,\27\2\u0125\u0123\3\2\2\2\u0126\u0129\3\2\2\2\u0127\u0125",
-                      "\3\2\2\2\u0127\u0128\3\2\2\2\u0128)\3\2\2\2\u0129\u0127\3\2\2\2\u012a",
-                      "\u012b\7\24\2\2\u012b\u0130\5$\23\2\u012c\u012d\t\3\2\2\u012d\u012f",
-                      "\5$\23\2\u012e\u012c\3\2\2\2\u012f\u0132\3\2\2\2\u0130\u012e\3\2\2",
-                      "\2\u0130\u0131\3\2\2\2\u0131\u0133\3\2\2\2\u0132\u0130\3\2\2\2\u0133",
-                      "\u0134\7\37\2\2\u0134+\3\2\2\2\u0135\u013b\5\6\4\2\u0136\u013b\5\b",
-                      "\5\2\u0137\u013b\5\n\6\2\u0138\u013b\5\f\7\2\u0139\u013b\5\60\31\2",
-                      "\u013a\u0135\3\2\2\2\u013a\u0136\3\2\2\2\u013a\u0137\3\2\2\2\u013a",
-                      "\u0138\3\2\2\2\u013a\u0139\3\2\2\2\u013b-\3\2\2\2\u013c\u0141\5\6\4",
-                      "\2\u013d\u0141\5\f\7\2\u013e\u0141\5\b\5\2\u013f\u0141\5\60\31\2\u0140",
-                      "\u013c\3\2\2\2\u0140\u013d\3\2\2\2\u0140\u013e\3\2\2\2\u0140\u013f",
-                      "\3\2\2\2\u0141/\3\2\2\2\u0142\u0143\t\4\2\2\u0143\61\3\2\2\2/\63:>",
-                      "@HPSZ_chmpx|\u0081\u0085\u008e\u0092\u009a\u00a0\u00a8\u00b1\u00b7",
-                      "\u00bf\u00c3\u00cb\u00d1\u00d9\u00dd\u00ea\u00f1\u00fa\u00fe\u0102",
-                      "\u0106\u010a\u010e\u011a\u011d\u0120\u0127\u0130\u013a\u0140"].join("")
+                      "\u00a5\u00aa\5\4\3\2\u00a6\u00a7\7\t\2\2\u00a7\u00a9\5\4\3\2\u00a8",
+                      "\u00a6\3\2\2\2\u00a9\u00ac\3\2\2\2\u00aa\u00a8\3\2\2\2\u00aa\u00ab",
+                      "\3\2\2\2\u00ab\23\3\2\2\2\u00ac\u00aa\3\2\2\2\u00ad\u00ae\7\24\2\2",
+                      "\u00ae\u00af\5$\23\2\u00af\u00b0\7\26\2\2\u00b0\u00b9\5\26\f\2\u00b1",
+                      "\u00b3\7\27\2\2\u00b2\u00b4\5$\23\2\u00b3\u00b2\3\2\2\2\u00b3\u00b4",
+                      "\3\2\2\2\u00b4\u00b5\3\2\2\2\u00b5\u00b6\7\26\2\2\u00b6\u00b8\5\26",
+                      "\f\2\u00b7\u00b1\3\2\2\2\u00b8\u00bb\3\2\2\2\u00b9\u00b7\3\2\2\2\u00b9",
+                      "\u00ba\3\2\2\2\u00ba\u00bc\3\2\2\2\u00bb\u00b9\3\2\2\2\u00bc\u00bd",
+                      "\7\37\2\2\u00bd\25\3\2\2\2\u00be\u00c0\5\30\r\2\u00bf\u00be\3\2\2\2",
+                      "\u00c0\u00c1\3\2\2\2\u00c1\u00bf\3\2\2\2\u00c1\u00c2\3\2\2\2\u00c2",
+                      "\27\3\2\2\2\u00c3\u00c6\5\f\7\2\u00c4\u00c6\5\24\13\2\u00c5\u00c3\3",
+                      "\2\2\2\u00c5\u00c4\3\2\2\2\u00c6\31\3\2\2\2\u00c7\u00c8\7\24\2\2\u00c8",
+                      "\u00c9\5$\23\2\u00c9\u00ca\7\26\2\2\u00ca\u00d3\5\34\17\2\u00cb\u00cd",
+                      "\7\27\2\2\u00cc\u00ce\5$\23\2\u00cd\u00cc\3\2\2\2\u00cd\u00ce\3\2\2",
+                      "\2\u00ce\u00cf\3\2\2\2\u00cf\u00d0\7\26\2\2\u00d0\u00d2\5\34\17\2\u00d1",
+                      "\u00cb\3\2\2\2\u00d2\u00d5\3\2\2\2\u00d3\u00d1\3\2\2\2\u00d3\u00d4",
+                      "\3\2\2\2\u00d4\u00d6\3\2\2\2\u00d5\u00d3\3\2\2\2\u00d6\u00d7\7\37\2",
+                      "\2\u00d7\33\3\2\2\2\u00d8\u00da\5\36\20\2\u00d9\u00d8\3\2\2\2\u00da",
+                      "\u00db\3\2\2\2\u00db\u00d9\3\2\2\2\u00db\u00dc\3\2\2\2\u00dc\35\3\2",
+                      "\2\2\u00dd\u00e0\5.\30\2\u00de\u00e0\5\32\16\2\u00df\u00dd\3\2\2\2",
+                      "\u00df\u00de\3\2\2\2\u00e0\37\3\2\2\2\u00e1\u00e2\7\24\2\2\u00e2\u00e3",
+                      "\5$\23\2\u00e3\u00f3\7\26\2\2\u00e4\u00ec\5\"\22\2\u00e5\u00e6\7\27",
+                      "\2\2\u00e6\u00e7\5$\23\2\u00e7\u00e8\7\26\2\2\u00e8\u00e9\5\"\22\2",
+                      "\u00e9\u00eb\3\2\2\2\u00ea\u00e5\3\2\2\2\u00eb\u00ee\3\2\2\2\u00ec",
+                      "\u00ea\3\2\2\2\u00ec\u00ed\3\2\2\2\u00ed\u00ef\3\2\2\2\u00ee\u00ec",
+                      "\3\2\2\2\u00ef\u00f0\7\27\2\2\u00f0\u00f1\7\26\2\2\u00f1\u00f2\5\"",
+                      "\22\2\u00f2\u00f4\3\2\2\2\u00f3\u00e4\3\2\2\2\u00f3\u00f4\3\2\2\2\u00f4",
+                      "\u00f5\3\2\2\2\u00f5\u00f6\7\37\2\2\u00f6!\3\2\2\2\u00f7\u00fc\5\16",
+                      "\b\2\u00f8\u00f9\7\7\2\2\u00f9\u00fb\5\16\b\2\u00fa\u00f8\3\2\2\2\u00fb",
+                      "\u00fe\3\2\2\2\u00fc\u00fa\3\2\2\2\u00fc\u00fd\3\2\2\2\u00fd#\3\2\2",
+                      "\2\u00fe\u00fc\3\2\2\2\u00ff\u0101\7\35\2\2\u0100\u00ff\3\2\2\2\u0100",
+                      "\u0101\3\2\2\2\u0101\u0104\3\2\2\2\u0102\u0105\5(\25\2\u0103\u0105",
+                      "\5*\26\2\u0104\u0102\3\2\2\2\u0104\u0103\3\2\2\2\u0105\u0110\3\2\2",
+                      "\2\u0106\u0108\t\3\2\2\u0107\u0109\7\35\2\2\u0108\u0107\3\2\2\2\u0108",
+                      "\u0109\3\2\2\2\u0109\u010c\3\2\2\2\u010a\u010d\5(\25\2\u010b\u010d",
+                      "\5*\26\2\u010c\u010a\3\2\2\2\u010c\u010b\3\2\2\2\u010d\u010f\3\2\2",
+                      "\2\u010e\u0106\3\2\2\2\u010f\u0112\3\2\2\2\u0110\u010e\3\2\2\2\u0110",
+                      "\u0111\3\2\2\2\u0111%\3\2\2\2\u0112\u0110\3\2\2\2\u0113\u011d\7\b\2",
+                      "\2\u0114\u011d\7\30\2\2\u0115\u0116\7\30\2\2\u0116\u011d\7\b\2\2\u0117",
+                      "\u011d\7\31\2\2\u0118\u0119\7\31\2\2\u0119\u011d\7\b\2\2\u011a\u011b",
+                      "\7\35\2\2\u011b\u011d\7\b\2\2\u011c\u0113\3\2\2\2\u011c\u0114\3\2\2",
+                      "\2\u011c\u0115\3\2\2\2\u011c\u0117\3\2\2\2\u011c\u0118\3\2\2\2\u011c",
+                      "\u011a\3\2\2\2\u011d\'\3\2\2\2\u011e\u0120\7\22\2\2\u011f\u011e\3\2",
+                      "\2\2\u011f\u0120\3\2\2\2\u0120\u0122\3\2\2\2\u0121\u0123\5&\24\2\u0122",
+                      "\u0121\3\2\2\2\u0122\u0123\3\2\2\2\u0123\u0124\3\2\2\2\u0124\u0129",
+                      "\5,\27\2\u0125\u0126\7\27\2\2\u0126\u0128\5,\27\2\u0127\u0125\3\2\2",
+                      "\2\u0128\u012b\3\2\2\2\u0129\u0127\3\2\2\2\u0129\u012a\3\2\2\2\u012a",
+                      ")\3\2\2\2\u012b\u0129\3\2\2\2\u012c\u012d\7\24\2\2\u012d\u0132\5$\23",
+                      "\2\u012e\u012f\t\3\2\2\u012f\u0131\5$\23\2\u0130\u012e\3\2\2\2\u0131",
+                      "\u0134\3\2\2\2\u0132\u0130\3\2\2\2\u0132\u0133\3\2\2\2\u0133\u0135",
+                      "\3\2\2\2\u0134\u0132\3\2\2\2\u0135\u0136\7\37\2\2\u0136+\3\2\2\2\u0137",
+                      "\u013d\5\6\4\2\u0138\u013d\5\b\5\2\u0139\u013d\5\n\6\2\u013a\u013d",
+                      "\5\f\7\2\u013b\u013d\5\60\31\2\u013c\u0137\3\2\2\2\u013c\u0138\3\2",
+                      "\2\2\u013c\u0139\3\2\2\2\u013c\u013a\3\2\2\2\u013c\u013b\3\2\2\2\u013d",
+                      "-\3\2\2\2\u013e\u0143\5\6\4\2\u013f\u0143\5\f\7\2\u0140\u0143\5\b\5",
+                      "\2\u0141\u0143\5\60\31\2\u0142\u013e\3\2\2\2\u0142\u013f\3\2\2\2\u0142",
+                      "\u0140\3\2\2\2\u0142\u0141\3\2\2\2\u0143/\3\2\2\2\u0144\u0145\t\4\2",
+                      "\2\u0145\61\3\2\2\2/\63:>@HPSZ_chmpx|\u0081\u0085\u008e\u0092\u009a",
+                      "\u00a0\u00aa\u00b3\u00b9\u00c1\u00c5\u00cd\u00d3\u00db\u00df\u00ec",
+                      "\u00f3\u00fc\u0100\u0104\u0108\u010c\u0110\u011c\u011f\u0122\u0129",
+                      "\u0132\u013c\u0142"].join("")
 
 	@@_ATN = Antlr4::Runtime::ATNDeserializer.new().deserialize(@@_serializedATN)