+++ spec/modl/modl3.rb
@@ -1,0 +1,65 @@
+#!/usr/bin/ruby
+# frozen_string_literal: true
+
+# The MIT License (MIT)
+#
+# Copyright (c) 2019 NUM Technology Ltd
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+
+
+require 'modl'
+
+def process_modl(a)
+  begin
+    str = MODL::Interpreter.interpret(a, false)
+    puts str
+    return str
+  rescue MODL::InterpreterError => e
+    puts e.to_s
+  rescue MODL::ParserError => e
+    puts e.to_s
+  end
+
+end
+
+a = [
+    "_abc=xyz;_def=pqr; {%`xyz`=%abc?    result=success /? result=fail}", # lhs explicitly literal
+    "_abc=xyz;_def=pqr; {abc=xyz?        result=success /? result=fail}", # rhs implicitly literal
+    "_abc=xyz;_def=pqr; {%`xyz`=%`xyz`?  result=success /? result=fail}", # lhs+rhs explicitly literal
+    "_abc=xyz;_def=pqr; {%`xyz`=xyz?     result=success /? result=fail}",
+    "_abc=xyz;_def=pqr; {abc=%`xyz`?     result=success /? result=fail}",
+    "_abc=xyz;_def=pqr; {abc=%`XYZ`.d?   result=success /? result=fail}",
+    "_abc=xyz;_def=pqr; {abc=%abc?       result=success /? result=fail}",
+
+    "_abc=xyz;_def=pqr; {\"abc\"=%abc?   result=fail    /? result=success}",
+    "_abc=xyz;_def=pqr; {'abc'=%abc?     result=fail    /? result=success}",
+    "_abc=xyz;_def=pqr; {`abc`=%abc?     result=fail    /? result=success}",
+
+    "_abc=xyz;_def=pqr; {abc=abc?        result=fail    /? result=success}", # rhs defaults to literal
+    "_abc=xyz;_def=pqr; {abc=%`abc`?     result=fail    /? result=success}",
+    "_abc=xyz;_def=pqr; {abc=`%abc`?     result=fail    /? result=success}",
+    "_abc=xyz;_def=pqr; {abc=`%xyz`?     result=fail    /? result=success}",
+
+    "{abc=xyz?                           result=fail    /? result=success}", # abc is undefined
+    "_abc=xyz; {abc=%def?                result=fail    /? result=success}", # def is undefined
+
+]
+
+a.each { |m| print m + "\t\t\t"; process_modl(m) }